+++ src/dolphin/os/OSExec.c
@@ -47,7 +47,7 @@ static int PackArgs(void* address, int argc, char** argv)
         cursor = boot_info + ((cursor - boot_info) & ~3);
         cursor -= (argument_count + 1) * sizeof(char*);
         list = (char**)cursor;
-        for (i = 0; i < (unsigned long)argument_count + 1; i++) {
+        for (i = 0; i < argument_count + 1; i++) {
             list[i] = argv[i];
         }
 
@@ -67,7 +67,9 @@ static void Run(void* entry_point)
     ((void (*)(void))entry_point)();
 }
 
-static void ReadDisc(void* address, int length, int offset)
+/* TODO: [breakthrough needed] 65.18519%; DVDCommandBlock and loop semantics
+ * agree, but frame size, argument scheduling, and branch placement still differ. */
+static void ReadDisc(void* address, signed long length, signed long offset)
 {
     DVDCommandBlock block;
 
@@ -84,6 +86,7 @@ static void Callback(long result, DVDCommandBlock* block)
     Prepared = 1;
 }
 
+/* TODO: [breakthrough needed] 86.562500%; macro-based address test and ABI agree, but MWCC prologue/load scheduling remains unresolved after donor-shaped trial. */
 void __OSGetExecParams(OSExecParams* params)
 {
     if ((unsigned long)OS_EXEC_PARAMS >= 0x80000000) {
@@ -93,6 +96,8 @@ void __OSGetExecParams(OSExecParams* params)
     }
 }
 
+/* TODO: [near miss] 89.244896%; donor CFG and widths agree; unsigned offset
+ * test is already ABI-correct, leaving cmp/register encodings and 8-byte residue. */
 static int GetApploaderPosition(void)
 {
     static long apploader_position;
